mail_archive/requirements.txt
django==4.2.11
djangorestframework==3.15.2
drf-spectacular==0.27.2
celery==5.3.6
redis==5.0.4
boto3==1.34.106
elasticsearch==8.13.0
pyotp==2.9.0
mysqlclient==2.2.4
PyJWT==2.8.0
PyMySQL==1.1.1
